test(ui): split TOML builder coverage
